connect/requirements.txt
asgiref==3.3.4
Django==3.2.4
django-cors-headers==3.7.0
django-user-visit==0.4.1
django-utils-six==2.0
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
ua-parser==0.10.0
user-agents==2.2.0
